Rename internal libextobjc files to avoid clashes with standalone libextobjc files
