+++ Mantle/extobjc/MTLEXTScope.m
@@ -1,5 +1,5 @@
 //
-//  EXTScope.m
+//  MTLEXTScope.m
 //  extobjc
 //
 //  Created by Justin Spahr-Summers on 2011-05-04.
@@ -7,7 +7,7 @@
 //  Released under the MIT license.
 //
 
-#import "EXTScope.h"
+#import "MTLEXTScope.h"
 
 void mtl_executeCleanupBlock (__strong mtl_cleanupBlock_t *block) {
     (*block)();
