+++ Mantle/extobjc/include/MTLEXTRuntimeExtensions.h
@@ -1,5 +1,5 @@
 //
-//  EXTRuntimeExtensions.h
+//  MTLEXTRuntimeExtensions.h
 //  extobjc
 //
 //  Created by Justin Spahr-Summers on 2011-03-05.
